:recycle: refactor: create hook register
