+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export { useCart } from './useCart'
 export { useFavorites } from './useFavorites'
 export { useProductsCart } from './useProductsCart'
 export { useProducts } from './useProducts'
+export { useRegisterUser } from './useRegisterUser'
